Remove Errs() method from collection stream
Keep all channel logic inside collection stream code
